// src/generated/models/Ope_simplesecurityactionsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Ope_simplesecurityactionsope_operation = {
  884680000: 'associate',
  884680001: 'disassociate'
} as const;
export type Ope_simplesecurityactionsope_operation = keyof typeof Ope_simplesecurityactionsope_operation;
export const Ope_simplesecurityactionsope_principletype = {
  884680000: 'systemuser',
  884680001: 'team'
} as const;
export type Ope_simplesecurityactionsope_principletype = keyof typeof Ope_simplesecurityactionsope_principletype;
export const Ope_simplesecurityactionsope_relatedtype = {
  884680000: 'role',
  884680001: 'team',
  884680002: 'columnsecurityprofile'
} as const;
export type Ope_simplesecurityactionsope_relatedtype = keyof typeof Ope_simplesecurityactionsope_relatedtype;
export const Ope_simplesecurityactionsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Ope_simplesecurityactionsstatecode = keyof typeof Ope_simplesecurityactionsstatecode;
export const Ope_simplesecurityactionsstatuscode = {
  1: 'Pending',
  2: 'Inactive',
  884680001: 'Success',
  884680002: 'Failed'
} as const;
export type Ope_simplesecurityactionsstatuscode = keyof typeof Ope_simplesecurityactionsstatuscode;

export interface Ope_simplesecurityactionsBase {
  importsequencenumber?: string;
  ope_errormessage?: string;
  ope_name: string;
  ope_operation?: Ope_simplesecurityactionsope_operation;
  "ope_PrincipalTeam@odata.bind"?: string;
  "ope_PrincipalUser@odata.bind"?: string;
  ope_principletype?: Ope_simplesecurityactionsope_principletype;
  ope_relatedprofile?: string;
  "ope_RelatedRole@odata.bind"?: string;
  "ope_RelatedTeam@odata.bind"?: string;
  ope_relatedtype?: Ope_simplesecurityactionsope_relatedtype;
  ope_simplesecurityactionid: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Ope_simplesecurityactionsstatecode;
  statuscode?: Ope_simplesecurityactionsstatuscode;
  timezoneruleversionnumber?: string;
  utcconversiontimezonecode?: string;
}

export interface Ope_simplesecurityactions extends Ope_simplesecurityactionsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ope_operationname?: string;
  ope_principalteamname?: string;
  ope_principalteamyominame?: string;
  ope_principalusername?: string;
  ope_principaluseryominame?: string;
  ope_principletypename?: string;
  ope_relatedrolename?: string;
  ope_relatedteamname?: string;
  ope_relatedteamyominame?: string;
  ope_relatedtypename?: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  ope_principalteam?: object;
  _ope_principalteam_value?: string;
  ope_principaluser?: object;
  _ope_principaluser_value?: string;
  ope_relatedrole?: object;
  _ope_relatedrole_value?: string;
  ope_relatedteam?: object;
  _ope_relatedteam_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}